test: update tools tests and add thinking parser tests
